simulation : - séparation code spécifique hard - avancement de l'implémentation de la simulation utilisant qemu (rcc et tim1)
	modified:   include/io/pwm.h
	modified:   qemu
	modified:   scripts/arm-elf-gdb-localhost1234
	modified:   scripts/gdb
	modified:   simu/ArmCm3.cxx
	modified:   simu/ArmCm3.h
	new file:   simu/ArmMem.h
	modified:   simu/ArmRcc.cxx
	modified:   simu/ArmRcc.h
	new file:   simu/ArmTimMotor.cxx
	new file:   simu/ArmTimMotor.h
	modified:   simu/CpuEmu.cxx
	modified:   simu/Makefile
	new file:   simu/Motor.cxx
	new file:   simu/Motor.h
	modified:   src/build.mk
	renamed:    src/io/encoders.c -> src/io/arm_cm3/encoders.c
	new file:   src/io/arm_cm3/pwm.c
	new file:   src/io/gcc_posix/encoders.c
	new file:   src/io/gcc_posix/pwm.c
	deleted:    src/io/pwm.c
	modified:   src/test/build.mk

+++ include/io/pwm.h
@@ -2,7 +2,7 @@
 #define PWM_H
 
 //! @file pwm.h
-//! @brief PWM
+//! @brief PWM interface
 //! @author Jean-Baptiste Trédez
 
 #include <stdint.h>
@@ -17,6 +17,6 @@
 //!
 //! @param num pwm id
 //! @param val new value
-void pwm_set(unsigned int num, uint32_t val, int dir);
+void pwm_set(unsigned int num, uint16_t val, int dir);
 
 #endif
